Fixed sidebar, added comments and created a way to add circles on the map
Almost done figuring out a way to add circles to the map. We will need just basic information about the crime and change the colors on the circle depending on the information. The location of the circle may be difficult, but if we can at least get the county of the crime, we should be okay I believe.